+++ exhibitera/changelog.md
@@ -19,6 +19,7 @@ _Exhibitera Hub and Apps must all be upgraded to version 6 at the same time._
 - Creating, editing, and setting definitions is now much easier when not using Hub
 - Easily access DMX lighting scenes when not using Hub
 - Use *Shift-click* to easily select multiple files in the file picker
+- Drag and drop files on the file picker to start an upload
 
 ### Custom Apps
 - New Basic App option turns a simple HTML file into an app without needing to implement the Exhibitera APIs